+++ SituationOperator/Startup.cs
@@ -28,6 +28,7 @@ using SituationOperator.SituationManagers;
 using SituationDatabase.Models;
 using StackExchange.Redis;
 using Moq;
+using SituationOperator.Helpers.SubscriptionConfig;
 
 namespace SituationOperator
 {
@@ -208,6 +209,21 @@ namespace SituationOperator
 
                 return new EquipmentHelper(equipmentProvider);
             });
+            #endregion
+
+            #region Subscription Configuration
+
+            if (!GetOptionalEnvironmentVariable<bool>(Configuration, "MOCK_SUBSCRIPTION_LOADER", false))
+            {
+                services.AddSingleton<ISubscriptionConfigProvider, SubscriptionConfigLoader>();
+            }
+            else
+            {
+                Console.WriteLine(
+                    "WARNING: SubscriptionConfigLoader is mocked and will return mocked values!");
+                services.AddSingleton<ISubscriptionConfigProvider, MockedSubscriptionConfigLoader>();
+            }
+
             #endregion
 
             #region SituationManagers