+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Endpoint for situations of a specific type
 - SkillDomains
 - Replaced MatchSituationsModel.MatchId with MatchSituationsModel.MatchData
+- SubscriptionConfigLoader
 
 ## [0.1.0] - 2020-06-02
 ### Basic implementation