+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replaced MatchSituationsModel.MatchId with MatchSituationsModel.MatchData
 - SubscriptionConfigLoader
 
+### Changed
+- Remove exchange and rename queue name AMQP_EXCHANGE_CONSUME_QUEUE to AMQP_EXTRACTION_INSTRUCTIONS
+- Remove exchange and rename queue name AMQP_DEMOCENTRAL_REPLY to AMQP_EXTRACTION_REPLY
+
 ## [0.1.0] - 2020-06-02
 ### Basic implementation